Optimizers

Algorithms to solve optimization problems. Currently the L-BFGS algorithm is implemented, including box constraints (optional).

Public API

InverseAlgos.Optimizers — Module

Optimizers

A module collecting a set of deterministic inversion algorithms. Main targets are geophysical inverse problems.

Exports

gaussnewton

lmbfgs

InverseAlgos.Optimizers.lmbfgs — Function

lmbfgs(f::Function, ∇f::Function, args...)

An implementation of the L-BFGS algorithm following Nocedal & Wright, 2006 with the addition of box constraints.

Arguments

f: a function returning the misfit (a ::Function)

∇f: a function returning the gradient (a ::Function)

x0: the starting model/initial guess

mem: the length (number of iterations) used for memory variables

maxiter: maximum number of iterations

bounds (optional): a two-column array where the first column contains lower bounds (constraints) and the second upper bounds

target_update (optional): initial step length for the line search

outfile (optional): the name of the output file where to save the results

τgrad (optional): minimum value of the gradient at which to stop the algorithm

overwriteoutput (optional): if true overwrite the output file if already existing

maxiterwolfe (optional): maximum number of iterations for the line search function

maxiterzoom (optional): maximum number of iterations for the zoom function

c1 and c2 (optional): strong Wolfe values

saveres (optional): save results? Defaults to true

Returns

x: a vector containing the solution for each iterarion

misf: a vector containing the misfit value for each iteration

lmbfgs(
    fh!::Function,
    x0::Vector{<:Real};
    bounds,
    mem,
    maxiter,
    target_update,
    outfile,
    τgrad,
    overwriteoutput,
    maxiterwolfe,
    maxiterzoom,
    c1,
    c2,
    saveres
)

An implementation of the L-BFGS algorithm following Nocedal & Wright, 2006 with the addition of box constraints.

Arguments

fh!: a function (::Function) returning the misfit to be minimized and computing its gradient in place, e.g., misf=fh!(grad,x)

x0: the starting model/initial guess

mem: the length (number of iterations) used for memory variables

maxiter: maximum number of iterations

bounds (optional): a two-column array where the first column contains lower bounds (constraints) and the second upper bounds

target_update (optional): initial step length for the line search

outfile (optional): the name of the output file where to save the results

τgrad (optional): minimum value of the gradient at which to stop the algorithm

overwriteoutput (optional): if true overwrite the output file if already existing

maxiterwolfe (optional): maximum number of iterations for the line search function

maxiterzoom (optional): maximum number of iterations for the zoom function

c1 and c2 (optional): strong Wolfe values

saveres (optional): save results? Defaults to true

Returns

x: a vector containing the solution for each iterarion

misf: a vector containing the misfit value for each iteration